Collection of sensor data and cloud uploading.
This app collects all the sensors data from a smartphone such as accelerometer, magnetometer, etc., and sending it to the cloud. This application makes use of AWS DynamoDB to store data in the cloud in a table format. Initailly, the app was created to collect the sensor data and send to the cloud where using those data detecting earthquake.